requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
requests
pillow
aiofiles
byteplus-sdk
openai